+++ src/Parser/DoctrineRst/IndexDirective.php
@@ -1,0 +1,35 @@
+<?php
+/**
+ * This file is part of Berlioz framework.
+ *
+ * @license   https://opensource.org/licenses/MIT MIT License
+ * @copyright 2020 Ronan GIRON
+ * @author    Ronan GIRON <https://github.com/ElGigi>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code, to the root.
+ */
+
+declare(strict_types=1);
+
+namespace Berlioz\DocParser\Parser\DoctrineRst;
+
+use Doctrine\RST\Directives\Directive;
+use Doctrine\RST\Nodes\Node;
+use Doctrine\RST\Parser;
+
+class IndexDirective extends Directive
+{
+    public function getName(): string
+    {
+        return 'index';
+    }
+
+    /**
+     * @param string[] $options
+     */
+    public function processNode(Parser $parser, string $variable, string $data, array $options): ?Node
+    {
+        return new IndexNode($options);
+    }
+}